requirements.txt
asgiref==3.8.1
Django==5.1.7
django-ckeditor==6.7.2
django-js-asset==3.1.2
pillow==11.1.0
sqlparse==0.4.2
tzdata==2022.1
